+++ Analyzer/src/AnalyzeGenStop.cc
@@ -121,7 +121,7 @@ void AnalyzeGenStop::Loop(NTupleReader& tr, double, int maxevents, bool)
             if( !passMadHT ) continue; //Make sure not to double count DY events
             // Define Lumi weight
             const auto& Weight  = tr.getVar<float>("Weight");
-            const auto& lumi = tr.getVar<float>("Lumi");
+            const auto& lumi = tr.getVar<double>("Lumi");
             eventweight = lumi*Weight;
             
             // Define lepton weight
@@ -132,10 +132,10 @@ void AnalyzeGenStop::Loop(NTupleReader& tr, double, int maxevents, bool)
                 leptonScaleFactor = (GoodLeptons[0].first == "e") ? eleLepWeight : muLepWeight;
             }
             
-            bTagScaleFactor   = tr.getVar<float>("bTagSF_EventWeightSimple_Central");
+            bTagScaleFactor   = tr.getVar<double>("bTagSF_EventWeightSimple_Central");
             htDerivedScaleFactor = tr.getVar<float>("htDerivedweight");
-            prefiringScaleFactor = tr.getVar<float>("prefiringScaleFactor");
-            puScaleFactor = tr.getVar<float>("puWeightCorr");
+            prefiringScaleFactor = tr.getVar<double>("prefiringScaleFactor");
+            puScaleFactor = tr.getVar<double>("puWeightCorr");
             
                     weight *= eventweight*leptonScaleFactor*bTagScaleFactor*htDerivedScaleFactor*prefiringScaleFactor*puScaleFactor;
         }
